healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc

Former-commit-id: 6336f09998aa21ade58b350eada81abe18b52c6a
Former-commit-id: 3d718f7aca2c7e542c3c3a577b21f1a8183c106f
Former-commit-id: d36809962eb105b1eb9df004d62598cf8b225f03
Former-commit-id: cb95328c26847e96be003c8d592b8adb281e0253
Former-commit-id: f64a0886413a247dc7aea1e84430ac252ef08931
Former-commit-id: 599f17775d7fe72a67005587a1143008ac3c0a1b
Former-commit-id: bbd1dacdaccebca7bdd710daae8d2f208a5421a8